renamed files/relocated shared pages, added testing page
