+++ src/app/app.module.ts
@@ -18,10 +18,10 @@ import { RecentReleaseComponent } from './shared/components/recent-release/recen
 import { FontAwesomeModule, FaIconLibrary } from '@fortawesome/angular-fontawesome';
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
 import { DownloadsComponent } from './downloads/downloads.component';
-import { ReferencesComponent } from './references/references.component';
+import { ReferencesComponent } from './shared/components/references/references.component';
 import { ReferenceItemComponent } from './shared/components/reference-item/reference-item.component';
 import { PaginationComponent } from './shared/components/pagination/pagination.component';
-import { ContributingComponent } from './contributing/contributing.component';
+import { ContributingComponent } from './shared/components/contributing/contributing.component';
 import { WeatherSimulationComponent } from './weather/weather-simulation/weather-simulation.component';
 import { WeatherSourcesComponent } from './weather/weather-sources/weather-sources.component';
 import { WeatherRegionComponent } from './weather/weather-region/weather-region.component';
@@ -29,7 +29,8 @@ import { WeatherCountryComponent } from './weather/weather-country/weather-count
 import { WeatherLocationComponent } from './weather/weather-location/weather-location.component';
 import { WeatherLocationUsCanComponent } from './weather/weather-location-us-can/weather-location-us-can.component';
 import { WeatherSearchComponent } from './weather/weather-search/weather-search.component';
-import { InstallationComponent } from './installation/installation.component';
+import { InstallationComponent } from './shared/components/installation/installation.component';
+import { TestingComponent } from './shared/components/testing/testing.component';
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 import { faArrowCircleRight } from '@fortawesome/free-solid-svg-icons';
 import { faBuilding } from '@fortawesome/free-regular-svg-icons';
@@ -68,6 +69,7 @@ import { faCaretDown } from '@fortawesome/free-solid-svg-icons';
     WeatherLocationUsCanComponent,
     WeatherSearchComponent,
     InstallationComponent,
+    TestingComponent,
   ],
   imports: [
     BrowserModule,
