--- src/app/contributing/contributing.component.ts
@@ -1,13 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-controbuting',
-  templateUrl: './contributing.component.html',
-  styleUrls: ['./contributing.component.scss']
-})
-export class ContributingComponent {
-
-  constructor() {
-  }
-
-}
