--- src/app/installation/installation.component.ts
@@ -1,14 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-installation',
-  templateUrl: './installation.component.html',
-  styleUrls: ['./installation.component.scss']
-})
-
-export class InstallationComponent {
-
-  constructor() {
-  }
-
-}
