--- src/app/reference/reference.component.html
@@ -1,76 +1,0 @@
-
-<section class="col-sm-12">
-	<div style="font-size:0.9rem;">
-		<h1 class="page-header" [innerHTML]="reference_item.title"></h1>
-		<div>
-			<section>
-				<div>
-					<div class="submitted">
-						<span><i>Submitted by admin on {{ day }}, {{ date }} - {{ time }}</i></span>
-					</div>
-					<div>
-						<div>
-							<p [innerHTML]="reference_item.description"></p>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.publication_date == null ? '' : 'Publication Date:' }}</strong></div>
-							<div>{{ reference_item.publication_date }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.research_org == null ? '' : 'Research Org:' }}</strong></div>
-							<div>{{ reference_item.research_org }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.doi == null ? '' : 'doi:' }}</strong></div>
-							<div>{{ reference_item.doi }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.osti_id == null ? '' : 'osti_id:' }}</strong></div>
-							<div>{{ reference_item.osti_id }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.url == null ? '' : 'url:' }}</strong></div>
-							<div>{{ reference_item.url }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.authors == null ? '' : 'Authors:' }}</strong></div>
-							<div>{{ reference_item.authors }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.subject == null ? '' : 'Subjects:' }}</strong></div>
-							<div>{{ reference_item.subject }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.resource_type == null ? '' : 'Resource Type:' }}</strong></div>
-							<div>{{ reference_item.resource_type }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.sponsoring_org == null ? '' : 'Sponsoring Org:' }}</strong></div>
-							<div>{{ reference_item.sponsoring_org }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.source_title == null ? '' : 'Source Title:' }}</strong></div>
-							<div>{{ reference_item.source_title }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.volume == null ? '' : 'Volume:' }}</strong></div>
-							<div>{{ reference_item.volume }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.issue == null ? '' : 'Issue:' }}</strong></div>
-							<div>{{ reference_item.issue }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.page_start == null ? '' : 'Page Start:' }}</strong></div>
-							<div>{{ reference_item.page_start }}</div>
-						</div>
-						<div>
-							<div><strong>{{ reference_item.page_end == null ? '' : 'Page End:' }}</strong></div>
-							<div>{{ reference_item.page_end }}</div>
-						</div>
-					</div>
-				</div>
-			</section>
-		</div>
-	</div>
-</section>
