--- src/app/references/references.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-references',
-  templateUrl: './references.component.html',
-  styleUrls: ['./references.component.scss']
-})
-
-export class ReferencesComponent {
-
- constructor() {}
-}
