+++ src/app/shared/components/contributing/contributing.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-controbuting',
+  templateUrl: './contributing.component.html',
+  styleUrls: ['./contributing.component.scss']
+})
+export class ContributingComponent {
+
+  constructor() {
+  }
+
+}
