+++ src/app/shared/components/installation/installation.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-installation',
+  templateUrl: './installation.component.html',
+  styleUrls: ['./installation.component.scss']
+})
+
+export class InstallationComponent {
+
+  constructor() {
+  }
+
+}
