+++ src/app/shared/components/references/reference/reference.component.scss
@@ -1,0 +1,10 @@
+.submitted {
+	color:gray;
+	margin-top:-10px;
+	padding-bottom: 12px;
+}
+
+.active {
+	color: #3276B1;
+	font-weight: bold;
+}
