+++ src/app/shared/components/references/reference/reference.component.ts
@@ -1,0 +1,40 @@
+import { Component, Input, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { ReferenceItem } from '../../../classes/constants';
+import { references } from '../../../classes/reference';
+
+@Component({
+	selector: 'app-reference',
+  templateUrl: './reference.component.html',
+  styleUrls: ['./reference.component.scss']
+})
+
+export class ReferenceTemplateComponent implements OnInit {
+
+	@Input() reference_item: ReferenceItem;
+	@Input() reference: any[] = references;
+	@Input() day: String;
+	@Input() date: string;
+	@Input() time: string;
+
+	constructor(private route: ActivatedRoute) {
+		var time = new Date();
+		this.day = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'][time.getDay()];
+		this.date = time.getMonth() + 1 + '/' + time.getDate() + '/' + time.getFullYear();
+		this.time = time.getHours() + ':' + ('0' + time.getMinutes()).slice(-2);
+
+		for (let reference of this.reference) {
+			if (reference['id'] == route.url['_value'][1].path) {
+				this.reference_item = reference;
+			}
+		}
+	}
+
+  ngOnInit(): void {
+    if (!this.reference_item) {
+      throw new Error('ReferenceTemplateComponent attribute "reference_item" is required');
+		}
+	}
+
+ }
+
