+++ src/app/shared/components/references/references.component.html
@@ -1,0 +1,17 @@
+<div class="container">
+  <header role="banner"></header>
+  <div class="row">
+    <section class="col-sm-12" style="font-size:0.9rem;">
+      <ol class="breadcrumb" style="margin-left:-15px;width:1125px;">
+        <li class="breadcrumb-item"><a routerlink="/documentation">Documentation</a></li>
+        <li class="breadcrumb-item active" aria-current="page">References</li>
+      </ol>
+      <h1 style="margin-left:-15px;">References</h1>
+      <div class="region region-content" style="margin-left:-15px;width:1125px;">
+        <section id="block-system-main">
+          <app-reference-item></app-reference-item>
+        </section>
+      </div>
+  </section>
+  </div>
+</div>
