+++ src/app/shared/components/testing/testing.component.html
@@ -1,0 +1,61 @@
+<div class="row">
+	<section class="col-sm-12">
+		<a id="main-content"></a>
+		<h1 class="page-header">Testing and Validation</h1>
+		<div class="region region-content">
+			<section id="block-system-main" class="block block-system clearfix">
+				<div id="node-16" class="node node-page clearfix" about="/testing" typeof="foaf:Document">
+					<span property="dc:title" content="Testing and Validation" class="rdf-meta element-hidden"></span>
+					<div class="content" style="font-size:0.9rem;">
+						<div class="field field-name-body field-type-text-with-summary field-label-hidden"><div class="field-items"><div class="field-item even" property="content:encoded"><p>An important ongoing part of EnergyPlus development is testing using industry standard methods as major builds are completed. The goal is to make EnergyPlus as bug-free as possible. Three major types of tests are<br>currently conducted:</p>
+							<ul>
+								<li>Analytical tests:
+									<ul>
+										<li>HVAC tests, based on ASHRAE Research Project 865</li>
+										<li>Building fabric tests, based on ASHRAE Research Project 1052</li>
+									</ul>
+								</li>
+								<li>Comparative tests:
+									<ul>
+										<li>ANSI/ASHRAE Standard 140-2011</li>
+										<li>International Energy Agency Solar Heating and Cooling Programme (IEA SHC) BESTest (Building Energy Simulation Test) methods not yet in Standard 140</li>
+										<li>EnergyPlus HVAC Component Comparative tests</li>
+										<li>EnergyPlus Global Heat Balance tests</li>
+									</ul>
+								</li>
+								<li>Release and executable tests</li>
+							</ul>
+							<p>The BESTest suites compare the results of multiple simulation programs for a series of load-related attributes. For more information on IEA SHC Task 34/Annex 43 (BESTest), visit the <a href="http://www.iea-shc.org/">IEA Solar Heating and Cooling Programme website</a>.</p>
+							<p>EnergyPlus testing reports are available for  many of these test suites.<br>As new releases are completed and as we conclude new tests, we will add the results to this page.<br>Click on the links below to view the reports for EnergyPlus version 8.2.0.</p>
+							<ul>
+								<li>ANSI/ASHRAE Standard 140-2011
+									<ul>
+										<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE140-Envelope-8.3.0-b45b06b780.pdf">Envelope</a></li>
+										<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE140-HVAC-CE100200-8.3.0-b45b06b780.pdf">HVAC CE100-200</a></li>
+										<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE140-HVAC-CE300545-8.3.0-b45b06b780.pdf">HVAC CE300-545</a></li>
+										<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE140-HVAC-HE100230-8.3.0-b45b06b780.pdf">HVAC HE100-230</a></li>
+										<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE140-HERS-8.3.0-b45b06b780.pdf">HERS (Section 7)</a></li>
+									</ul>
+								</li>
+								<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE1052RP-8.3.0-b45b06b780.pdf">ASHRAE Research Project 1052</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/HVACComponent-8.3.0-b45b06b780.pdf">EnergyPlus HVAC Component Comparative Tests</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/GlobalEnergyBalance-8.3.0-b45b06b780.pdf">EnergyPlus Global Energy Balance Tests</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/ASHRAE140-GrndCoup-8.3.0-b45b06b780.pdf">IEA BESTEST In-Depth Ground Coupled Heat Transfer Tests</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/MultizoneBESTEST-8.3.0-b45b06b780.pdf">IEA BESTEST Multi-Zone Non-Airflow In-Depth Tests</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/current_testing_reports/IEAMechanical-8.3.0-b45b06b780.pdf">IEA BESTEST Mechanical Equipment &amp; Control Strategies for a Chilled Water and a Hot Water System</a></li>
+							</ul><p>Testing reports for previous versions of EnergyPlus are available below:</p>
+							<ul>
+								<li><a href="assets/nrel_custom/eplus_files/archived_testing_reports/energyplus_7.0.0.036_testing_reports.zip">EnergyPlus 7.0.0.036 Testing Reports</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/archived_testing_reports/energyplus_7.1.0.012_testing_reports.zip">EnergyPlus 7.1.0.012 Testing Reports</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/archived_testing_reports/energyplus_7.2.0.006_testing_reports.zip">EnergyPlus 7.2.0.006 Testing Reports</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/archived_testing_reports/energyplus_8.0.0.008_testing_reports.zip">EnergyPlus 8.0.0.008 Testing Reports</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/archived_testing_reports/energyplus_8.1.0.009_testing_reports.zip">EnergyPlus 8.1.0.009 Testing Reports</a></li>
+								<li><a href="assets/nrel_custom/eplus_files/archived_testing_reports/energyplus_8.2.0.e53a815_testing_reports.zip">EnergyPlus 8.2.0.e53a815 Testing Reports</a></li>
+							</ul>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
+	</section>
+</div>
