+++ src/app/shared/components/testing/testing.component.scss
@@ -1,0 +1,4 @@
+a {
+	color: #3276B1;
+	font-weight: bold;
+}
