+++ src/app/shared/components/testing/testing.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-testing',
+  templateUrl: './testing.component.html',
+  styleUrls: ['./testing.component.scss']
+})
+export class TestingComponent {
+
+  constructor() {
+  }
+}
